backend/requirements.txt
black==21.7b0
flake8==3.9.2
flask==2.0.1
gunicorn==20.1.0
keras==2.6.0
matplotlib==3.4.3
pillow==8.3.1
scikit_image==0.18
tensorflow==2.6.0
